+++ src/mx_memrchr.c
@@ -3,7 +3,7 @@
 void *mx_memrchr(const void *s, int c, size_t n) {
     char *temp = (char*) s;
     
-    for (unsigned long i = n - 1 ; i >= 0; i--)
+    for (size_t i = n - 1 ; i >= 0; i--)
         if (temp[i] == c)
             return (void*) &temp[i];
